refactor(server): move Handler type to _shared.ts
All route files imported Handler from api-routes.ts while
api-routes.ts imported from those same files — a circular
type import. Moving the type to _shared.ts (which routes
already use) breaks the cycle. api-routes.ts re-exports
Handler for channel-routes.ts and any external consumers.
