+++ src/server/mcp/routes/open.ts
@@ -1,6 +1,6 @@
 import type { Request, Response } from "express";
-import type { Handler } from "../api-routes.js";
 import { openFileByPath } from "../file-opener.js";
+import type { Handler } from "./_shared.js";
 import { sendApiError } from "./_shared.js";
 
 export function makeOpenHandler(): Handler {
