+++ src/server/mcp/routes/upload.ts
@@ -1,7 +1,7 @@
 import type { Request, Response } from "express";
-import type { Handler } from "../api-routes.js";
 import { detectFormat } from "../document-model.js";
 import { openFileFromContent } from "../file-opener.js";
+import type { Handler } from "./_shared.js";
 import { sendApiError } from "./_shared.js";
 
 export function makeUploadHandler(): Handler {
